requirements.txt
torch>=2.5.0
torchvision>=0.20.0
transformers>=4.46.2
huggingface-hub>=0.25.1
sentencepiece>=0.2.0
jinja2>=3.1.4
pydantic>=2.9.2
timm>=1.0.9
tiktoken>=0.8.0
numpy==1.26.4
accelerate>=1.1.1
sentence_transformers>=3.1.1
gradio>=5.6.0
openai>=1.55.0
einops>=0.8.0
pillow>=10.4.0
sse-starlette>=2.1.3
bitsandbytes>=0.44.2 # INT4 Loading

# vllm>=0.6.4.post1 # using with VLLM Framework
# peft>=0.14.0 # Using with finetune model
